BE/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI & Agents (chatbot_service)
langchain==0.2.5
langgraph==0.1.15
langchain-google-genai==1.0.6
google-generativeai==0.5.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2